SocialUrlMatabots.html
I deleted the pinterest image from the footer. I added links to the ones I found which where facebook and twitter and Instagram, I left a link to the home matabot page for the rest. I didn't connect it to the css since the images already had the css styles.